tiltfile: build_args should deserialize blobs to strings
